config/pyproject.toml
[tool.poetry]
name = "config"
version = "0.1.0"
description = ""
authors = ["IcaroOm <icaroomdigital@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.2"
asgiref = "3.6.0"
backports-zoneinfo = "0.2.1"
django = "4.2"
django-filter = "23.1"
djangorestframework = "3.14.0"
exceptiongroup = "1.1.1"
iniconfig = "2.0.0"
packaging = "23.1"
pluggy = "1.0.0"
pytest = "7.3.1"
pytz = "2023.3"
sqlparse = "0.4.4"
tomli = "2.0.1"
pytest-django = "^4.5.2"
blue = "^0.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
